server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.9.2
scikit-learn==1.5.2
python-multipart==0.0.9
python-dotenv==1.0.0
aiosqlite==0.20.0
pytest==7.4.0
pytest-asyncio==0.21.0
PyYAML==6.0.1